Separate between plotting and printing statistics summaries.
